% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canons.R
\docType{data}
\name{canons}
\alias{canons}
\title{canons: canonical miRNA expression data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{id}{Unique identifier for each miRNA sequence (character).}
  \item{mir_names}{Collapsed miRNA names corresponding to each `id` (character).}
  \item{Read}{The sequence read corresponding to the miRNA (character).}
  \item{tissue_sample}{Name of the tissue and sample (character).}
  \item{rpm}{Expression value (Reads Per Million) for the miRNA in the corresponding tissue sample (numeric).}
  \item{system}{Tissue system associated with the tissue (factor).}
  \item{tissue}{Specific tissue from which the sample was taken (factor).}
  \item{new_lab}{Label for the tissue used in analysis (factor).}
  \item{source}{Source of the sample (factor).}
  \item{breed}{Breed of the horse from which the tissue sample was obtained (factor).}
  \item{sex}{Sex of the horse (factor).}
}
}
\source{
The data was originally sourced from the `isomir_rpm.isomirmap_exp_filt.tsv` file. The miRNA expression 
values were normalized as Reads Per Million (RPM) to allow for comparison across samples.
}
\usage{
canons
}
\description{
This dataset contains canonical miRNA expression values across multiple tissue samples.
The miRNAs included are canonical (referred to as "canon" in the `Variant` column).
}
\details{
This dataset was generated by filtering and organizing miRNA data to retain only canonical variants, 
replacing missing values in the `Variant` column with "canon". The dataset was then transformed to a 
long format to facilitate analysis across different tissue samples. 

The `tissue_sample` column was used to join this data with additional metadata, providing contextual 
information such as the tissue system, tissue type, and sample source.
}
\examples{
# Loading the dataset
data(canons)

# Summarizing the dataset
summary(canons)

# Plotting RPM values for a specific miRNA across different tissues
library(ggplot2)
ggplot(canons, aes(x = tissue_sample, y = rpm, color = tissue)) +
  geom_point()

}
\keyword{datasets}
