% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_cts.R
\docType{data}
\name{proc_cts}
\alias{proc_cts}
\title{proc_cts: processed counts dataset}
\format{
A data frame with the following columns:
\describe{
  \item{tissue_sample}{Name of the tissue and sample (character).}
  \item{system}{Tissue system associated with the tissue (factor).}
  \item{tissue}{Specific tissue from which the sample was taken (factor).}
  \item{abvs}{Abbreviation for the tissue type.}
  \item{sample}{Sample identifier.}
  \item{breed}{Breed of the horse from which the tissue sample was obtained (factor).}
  \item{sex}{Sex of the horse (factor).}
  \item{source}{Source of the sample (factor).}
  \item{new_lab}{Label for the tissue used in analysis (factor).}
  \item{step}{The processing step (e.g., library_size, adapters_removed, quality_seqs_min_length, etc.).}
  \item{count}{Numeric value representing the read count associated with the respective processing step for each sample.}
  \item{source_mod}{Source with non-primary and FAANG samples as public (factor).}
}
}
\source{
Data was prepared using the script provided in `data-raw/process_fastqs_counts.csv` and joined with the `tissues` dataset.
}
\usage{
proc_cts
}
\description{
This dataset contains information about various processing steps applied to tissue samples, including library size, adapters removed, quality sequences, and more.
}
\examples{
# Example of loading and inspecting the proc_cts dataset
data(proc_cts)
head(proc_cts)
}
\keyword{datasets}
