% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid_rpms.R
\docType{data}
\name{uid_rpms}
\alias{uid_rpms}
\title{uid_rpms: unique identifier RPMs data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{id}{Unique identifier for each miRNA sequence (character).}
  \item{parents}{Collapsed parent miRNA names associated with each `id` (character).}
  \item{mir_names}{Collapsed miRNA names corresponding to each `id` (character).}
  \item{variants}{Collapsed variant names corresponding to each `id` (character).}
  \item{type}{Type of miRNA, either "canon" for canonical miRNAs or "isomiR" for isomiRs (factor).}
  \item{rpm}{Expression value (Reads Per Million) for the miRNA in the corresponding tissue sample (numeric).}
  \item{sample}{Unique identifier for the specific sample (factor).}
  \item{system}{Tissue system associated with the tissue (factor).}
  \item{tissue}{Specific tissue from which the sample was taken (factor).}
  \item{new_lab}{Label for the tissue used in analysis (factor).}
  \item{source}{Source of the sample (factor).}
  \item{breed}{Breed of the horse from which the tissue sample was obtained (factor).}
  \item{sex}{Sex of the horse (factor).}
}
}
\source{
The data was originally sourced from the `isomir_rpm.isomirmap_exp_filt.tsv` file.
}
\usage{
uid_rpms
}
\description{
This dataset contains miRNA expression values (Reads Per Million, RPM) organized by unique identifiers (UIDs) across various tissue samples. The data includes both canonical miRNAs and isomiRs, with additional contextual information about the miRNAs and their variants.
}
\details{
This dataset was generated by filtering miRNA data to include both canonical miRNAs and isomiRs. The data was transformed to collapse multiple entries for the same `id` into single entries, with the parent names, variant names, and miRNA names combined into single fields.

The `type` column differentiates between canonical miRNAs and isomiRs, allowing for specific analysis of these two types. The dataset was then pivoted to a long format for easy comparison across tissue samples.
}
\examples{
# Loading the dataset
data(uid_rpms)

# Summarizing the dataset
summary(uid_rpms)
}
\keyword{datasets}
